Moved redirectOnFareType functionality to defineTimeRestrictions
